Add safety delete

+++ utils/callbacks.py
@@ -1,10 +1,15 @@
 from aiogram import Router, F
 from aiogram.types import CallbackQuery
 
+from config import settings
+
 router = Router()
 
 @router.callback_query(F.data == "delete_message")
 async def delete_message_callback(callback: CallbackQuery):
+    is_admin = callback.from_user.id in settings.ADMIN_IDS
+    if not is_admin:
+        return
     try:
         await callback.message.delete()
     except Exception:
